+++ src/app/locales/en/components/-browser-columns-name/archive-filesystem.js
@@ -1,4 +1,4 @@
-import headers from '../-browser-columns-name/-browser-columns';
+import headers from './filesystem';
 
 export default {
   headers,