+++ src/app/locales/en/components/columns-configuration-popover/column-item.js
@@ -1,7 +1,7 @@
 import archiveFilesystem from '../-browser-columns-name/archive-filesystem';
 import archive from '../-browser-columns-name/archive';
 import dataset from '../-browser-columns-name/dataset';
-import filesystem from '../-browser-columns-name/-browser-columns';
+import filesystem from '../-browser-columns-name/filesystem';
 import shareFilesystem from '../-browser-columns-name/share-filesystem';
 import transfer from '../-browser-columns-name/transfer';
 
@@ -13,6 +13,6 @@ export default {
   archive,
   dataset,
   filesystem,
-  shareFilesystem,
+  shareFilesystem: shareFilesystem.headers,
   transfer,
 };